add a sh file for calling the fetch
I tried many ways to get the fetch call to background without staying
attached to the current subshell. Ends up it's not possible to nohup or
fully disown a function that is defined inside the same file that's
calling it.

So simple answer, I've created a file `fetch.sh` that can be called
inside nohup. It will source `git-base.sh` and call the fetch function.