Remove unnecessary subshells from bash prompt
In the zsh prompt we need to use '%s' on the printf to prevent the
`${fg_bold[<color>]}` expressions from expanding before the colors have
been loaded in.

In the bash prompt this is unnecessary as the \\033 codes are base
control characters, and so have no dependency they rely on.