Bypass PS1s conversion of \[
In PS1 you need to escape non-printing characters, like the color codes.
The standard way is wrapping it in `\[` and `\]`. But for a dynamic
prompt, i.e. one that renders the results of a function every time the
prompt renders, that `\[` will be output as literals.

To fix this we bypass the conversion and wrap our non-printing
characters in the desired characters directly:

`\[` -> `\x01`
`\]` -> `\x02`