%Z returns the modified date, %Y returns the changed date
The OSX version of stat uses %m to return the 'modified' date since
epoch. The GNU version of stat doesn't have the %m format code.

%Z which was being used in @hallzy's fix returns the 'changed' date
since epoch, which only changes when you modify the metadata on a file.
This meant that when we `touch $dot_git/.lastupdatetime` the 'changed'
date doesn't change.

I've switched to the %Y flag which returns the 'modified' time. This is
the time that is changed when `touch` modifies the file.