Made changes recommeneded by @michaeldfallen
in fetch.sh "if" now checks for $@ instead of just $1

in radar-base.sh there is now a parameter expansion on line 170 instead of the
if statements.

in git-radar, now there is a shift, and a regex check that the next value is a
number. If the next value after --fetch_t is not a number, an error is echo'd
and it resorts to the default behaviour of 5 minutes.